Give the article's table its own module, and share the CSV export
The article panel was half table code. Moving it out also gives the comparison
panel somewhere to share the CSV export from: both had the same callback, and
neither needs to hold the exported rows in render state to produce it.
